Add possibility to force the table name
Add a vararg parameter to TABLE_INIT that allows to force to use a given
table name. So if we use TABLE_INIT(Name) the table name for the object
`Name` will be `names` but if we use TABLE_INIT(Name, my_table_name) the
table name for object `Name` will be `my_table_name`.